Split up adoptChild shim into smaller variants
