+++ Sources/Task/TaskProgress.swift
@@ -8,224 +8,11 @@
 
 import Foundation
 #if SWIFT_PACKAGE
-import Atomics
 import Deferred
-#elseif COCOAPODS
-import Atomics
-#elseif XCODE && !FORCE_PLAYGROUND_COMPATIBILITY
-import Deferred.Atomics
 #endif
 
 #if os(macOS) || os(iOS) || os(tvOS) || os(watchOS)
-/// A progress object whose attributes reflect that of an external progress
-/// tree.
-private final class ProxyProgress: Progress {
-
-    @objc dynamic let observee: Progress
-    lazy var token = Observation(forUpdating: self)
-
-    init(attachingTo observee: Progress) {
-        self.observee = observee
-        super.init(parent: .current())
-        totalUnitCount = 100
-        token.activate(observing: observee)
-    }
-
-    deinit {
-        token.invalidate(observing: observee)
-    }
-
-    // MARK: - Derived values
-
-    override func cancel() {
-        observee.cancel()
-    }
-
-    override func pause() {
-        observee.pause()
-    }
-
-    @available(macOS 10.11, iOS 9.0, watchOS 2.0, tvOS 9.0, *)
-    override func resume() {
-        observee.resume()
-    }
-
-    @objc static let keyPathsForValuesAffectingUserInfo: Set<String> = [
-        #keyPath(observee.userInfo)
-    ]
-
-    override var userInfo: [ProgressUserInfoKey: Any] {
-        return observee.userInfo
-    }
-
-    override func setUserInfoObject(_ objectOrNil: Any?, forKey key: ProgressUserInfoKey) {
-        observee.setUserInfoObject(objectOrNil, forKey: key)
-    }
-
-    // MARK: - KVO babysitting
-
-    func inheritFraction() {
-        completedUnitCount = observee.isIndeterminate ? -1 : Int64(observee.fractionCompleted * Double(totalUnitCount))
-    }
-
-    func inheritAttribute(_ value: Any?, forKeyPath keyPath: String) {
-        setValue(value, forKeyPath: keyPath)
-    }
-
-    func inheritCancelled() {
-        if observee.isCancelled {
-            super.cancel()
-        }
-    }
-
-    func inheritPaused() {
-        if observee.isPaused {
-            super.pause()
-        } else if #available(macOS 10.11, iOS 9.0, watchOS 2.0, tvOS 9.0, *) {
-            super.resume()
-        }
-    }
-
-    /// A side-table object to weakify the progress observer and prevent
-    /// delivery of notifications after deinit.
-    final class Observation: NSObject {
-        static var fractionContext = false
-        static var attributesContext = false
-        static var cancelledContext = false
-        static var pausedContext = false
-
-        static let fractionKeyPaths = [
-            #keyPath(Progress.fractionCompleted),
-            #keyPath(Progress.isIndeterminate)
-        ]
-
-        static let attributesKeyPaths = [
-            #keyPath(Progress.localizedDescription),
-            #keyPath(Progress.localizedAdditionalDescription),
-            #keyPath(Progress.cancellable),
-            #keyPath(Progress.pausable),
-            #keyPath(Progress.kind)
-        ]
-
-        struct State: OptionSet {
-            let rawValue: UInt8
-            static let ready = State(rawValue: 1 << 0)
-            static let observing = State(rawValue: 1 << 1)
-            static let cancellable: State = [.ready, .observing]
-            static let cancelled = State(rawValue: 1 << 2)
-        }
-
-        var state = UInt8() // see State
-        weak var observer: ProxyProgress?
-
-        init(forUpdating observer: ProxyProgress) {
-            self.observer = observer
-            bnr_atomic_init(&state, State.ready.rawValue)
-            super.init()
-        }
-
-        func activate(observing observee: Progress) {
-            for key in Observation.fractionKeyPaths {
-                observee.addObserver(self, forKeyPath: key, options: .initial, context: &Observation.fractionContext)
-            }
-
-            for key in Observation.attributesKeyPaths {
-                observee.addObserver(self, forKeyPath: key, options: [ .initial, .new ], context: &Observation.attributesContext)
-            }
-            observee.addObserver(self, forKeyPath: #keyPath(Progress.cancelled), options: .initial, context: &Observation.cancelledContext)
-            observee.addObserver(self, forKeyPath: #keyPath(Progress.paused), options: .initial, context: &Observation.pausedContext)
-
-            bnr_atomic_fetch_or(&state, State.observing.rawValue, .release)
-        }
-
-        func invalidate(observing observee: Progress) {
-            let oldState = State(rawValue: bnr_atomic_fetch_and(&state, ~State.ready.rawValue, .relaxed))
-            guard !oldState.isStrictSuperset(of: .cancellable) else { return }
-            bnr_atomic_fetch_or(&state, State.cancelled.rawValue, .relaxed)
-
-            for key in Observation.fractionKeyPaths {
-                observee.removeObserver(self, forKeyPath: key, context: &Observation.fractionContext)
-            }
-
-            for key in Observation.attributesKeyPaths {
-                observee.removeObserver(self, forKeyPath: key, context: &Observation.attributesContext)
-            }
-            observee.removeObserver(self, forKeyPath: #keyPath(Progress.cancelled), context: &Observation.cancelledContext)
-            observee.removeObserver(self, forKeyPath: #keyPath(Progress.paused), context: &Observation.pausedContext)
-        }
-
-        override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey: Any]?, context: UnsafeMutableRawPointer?) {
-            let state = State(rawValue: bnr_atomic_load(&self.state, .relaxed))
-            guard state.contains(.ready), let observer = observer else { return }
-            // This would be prettier as a switch.
-            // https://bugs.swift.org/browse/SR-7877
-            if context == &Observation.fractionContext {
-                observer.inheritFraction()
-            } else if context == &Observation.attributesContext, let keyPath = keyPath {
-                observer.inheritAttribute(change?[.newKey], forKeyPath: keyPath)
-            } else if context == &Observation.cancelledContext {
-                observer.inheritCancelled()
-            } else if context == &Observation.pausedContext {
-                observer.inheritPaused()
-            }
-        }
-    }
-}
-
 extension Progress {
-    /// Attempt a backwards-compatible implementation of iOS 9's explicit
-    /// progress handling. It's not perfect; this is a best effort of proxying
-    /// an external progress tree.
-    ///
-    /// If `progress` may possibly already have a parent,
-    /// send `orphaned: false`, using similar behavior to the backwards-
-    /// compatible path.
-    @discardableResult
-    func adoptChild(_ progress: Progress, orphaned canAdopt: Bool, pendingUnitCount: Int64) -> Progress {
-        if #available(macOS 10.11, iOS 9.0, watchOS 2.0, tvOS 9.0, *), canAdopt {
-            addChild(progress, withPendingUnitCount: pendingUnitCount)
-            return progress
-        } else {
-            let changedPendingUnitCount = Progress.current() === self
-            if changedPendingUnitCount {
-                resignCurrent()
-            }
-
-            becomeCurrent(withPendingUnitCount: pendingUnitCount)
-
-            let progress = ProxyProgress(attachingTo: progress)
-
-            if !changedPendingUnitCount {
-                resignCurrent()
-            }
-
-            return progress
-        }
-    }
-}
-
-// MARK: - Convenience initializers
-
-extension Progress {
-    /// A simple indeterminate progress with a cancellation function.
-    static func wrappingCompletion<Wrapped: FutureProtocol>(of wrapped: Wrapped, uponCancel cancellation: (() -> Void)?) -> Progress {
-        let totalUnitCount: Int64 = wrapped.peek() != nil ? 0 : -1
-        let progress = Progress(totalUnitCount: totalUnitCount)
-
-        if let cancellation = cancellation {
-            progress.cancellationHandler = cancellation
-        } else {
-            progress.isCancellable = false
-        }
-
-        wrapped.upon(.global(qos: .utility)) { _ in
-            progress.totalUnitCount = 1
-            progress.completedUnitCount = 1
-        }
-
-        return progress
-    }
-
     /// A simple indeterminate progress with a cancellation function.
     static func wrappingSuccess<Wrapped: TaskProtocol>(of wrapped: Wrapped, uponCancel cancellation: (() -> Void)? = nil) -> Progress {
         switch (wrapped as? Task<Wrapped.SuccessValue>, cancellation) {
@@ -234,10 +21,10 @@ extension Progress {
         case (let task?, let cancellation?):
             let progress = Progress(totalUnitCount: 1)
             progress.cancellationHandler = cancellation
-            progress.adoptChild(task.progress, orphaned: false, pendingUnitCount: 1)
+            progress.monitorChild(task.progress, withPendingUnitCount: 1)
             return progress
         default:
-            return .wrappingCompletion(of: wrapped, uponCancel: cancellation)
+            return .basicProgress(parent: .current(), for: wrapped, uponCancel: cancellation)
         }
     }
 }
@@ -269,7 +56,7 @@ extension Progress {
             // Otherwise, wrap it up as a Task<Value>-marked progress.
             let outer = Progress(totalUnitCount: taskRootUnitCount)
             outer.setUserInfoObject(NSLock(), forKey: taskRootLock)
-            outer.adoptChild(inner, orphaned: true, pendingUnitCount: taskRootUnitCount)
+            outer.adoptChild(inner, withPendingUnitCount: taskRootUnitCount)
             return outer
         }
     }
@@ -292,10 +79,9 @@ extension TaskProtocol {
         } else {
             let outer = Progress(totalUnitCount: taskRootUnitCount + continuedWorkUnitCount)
             outer.setUserInfoObject(NSLock(), forKey: taskRootLock)
-            outer.adoptChild(inner, orphaned: false, pendingUnitCount: taskRootUnitCount)
+            outer.monitorChild(inner, withPendingUnitCount: taskRootUnitCount)
             return outer
         }
     }
 }
-
 #endif
